[feat/#111]:error boundary 처리
